requirements.txt
altgraph==0.17.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
chatbotAI==0.3.1.3
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
construct==2.10.68
cx-Freeze==6.10
cx-Logging==3.0
decorator==5.1.1
deprecation==2.1.0
docopt==0.6.2
Eel==0.14.0
ffmpeg==1.4
future==0.17.1
fuzzywuzzy==0.18.0
geocoder==1.38.1
geojson==2.5.0
gevent==21.12.0
gevent-websocket==0.10.1
gmail==0.6.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
lief==0.12.1
loguru==0.6.0
lxml==4.8.0
notifypy==1.0.3.0
num2words==0.5.10
numpy==1.22.4
packaging==21.3
pefile==2022.5.30
Pillow==9.1.1
pvporcupine==1.9.5
pyaml==19.12.0
PyAudio @ file:///C:/1Coding/Python/S4M-2.0/conf/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycodestyle==2.8.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.14.1
pygame==2.1.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
PyJWT==2.4.0
pykeepass==4.0.2
pynput==1.7.6
pyowm==3.3.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-crfsuite==0.9.8
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
ratelim==0.1.6
requests==2.27.1
scikit-learn==0.22.2.post1
scipy==1.8.1
six==1.16.0
sklearn-crfsuite==0.3.6
snips-nlu==0.20.2
snips-nlu-de==0.2.3
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
sounddevice==0.4.4
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
tabulate==0.8.9
text2numde==1.0
tinydb==4.7.0
toml==0.10.2
tqdm==4.64.0
twilio==7.9.1
urllib3==1.26.9
vosk==0.3.7
whichcraft==0.6.1
wikipedia==1.4.0
win32-setctime==1.1.0
wxPython==4.1.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
